Added Intepreter
The interpreter interprets the output of the cv model (aka the most likely word) and gathers a list of them (e.g.) [the, the, the, quick, quick, brown, fox] and inputs it into ChatGPT. ChatGPT then turns that into a list of words that has no duplicates (e.g.) [the, quick, brown, fox] which will be much easier to turn into a real sentence later. It is encouraged to read the actual code to understand.